fixed multitasking demonstration
logic was wrong in several places.
fixed important compiler warnings and bad printf strings.
added 'make tests'
improved error handling. stopped ignoring error returns in some places.
pass errors up to the top program so that 'make tests' can fail.
kernel.c:made logging macros more consistent with rest of project.
task.c:added some tracing to track down problems
test2.c:added a second example to make it easier to tell programs apart.

+++ stackvm/kernel/prioq.h
@@ -3,7 +3,7 @@
 
 struct prioq_elm
 {
-    unsigned long d; /* key - usually time */
+    unsigned long long d; /* key - usually time */
     void *p;
 };
